+++ src/index.html
@@ -23,7 +23,7 @@
                 <hr>
                 <form action="" class="form-user" id="form-user">
                     <div class="form__group">
-                        <label for="">Username:</label>
+                        <label for="addUsername">Username:</label>
                         <input type="text" id="addUsername" placeholder="your name for example.." autocomplete="off">
                     </div>
                     <hr>
@@ -69,7 +69,7 @@
                 <div class="chat-action" id="chat-action"></div>
                 <form class="chat-form form-data">
                     <div class="form__group">
-                        <label aria-disabled="true">Username</label>
+                        <label aria-disabled="true" for="chat-username">Username</label>
                         <div class="fg--username">
                             <input type="text" placeholder="username" id="chat-username" readonly>
                             <button type="button" id="change-username" title="cambiar nombre de usuario">
